move ReqRep handling from controller to moonnode
